+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 <!-- _No unreleased changes_ -->
+### Added 
+- Added run 90 second ad button
 
 ## [1.1.1] - 2020-08-28
 ### Changed